fix template casting for Vector2i - float ops
Fixes https://github.com/wjakob/nanogui/issues/225

+++ src/screen.cpp
@@ -376,8 +376,8 @@ void Screen::drawWidgets() {
     glfwGetWindowSize(mGLFWWindow, &mSize[0], &mSize[1]);
 
 #if defined(_WIN32) || defined(__linux__)
-    mSize = (mSize / mPixelRatio).cast<int>();
-    mFBSize = (mSize * mPixelRatio).cast<int>();
+    mSize = (mSize.template cast<float>() / mPixelRatio).template cast<int>();
+    mFBSize = (mSize.template cast<float>() * mPixelRatio).template cast<int>();
 #else
     /* Recompute pixel ratio on OSX */
     if (mSize[0])
